Moving imports and documentation of IPCA to non-experimental
