+++ python/cuml/decomposition/__init__.py
@@ -14,6 +14,6 @@
 # limitations under the License.
 #
 
-from cuml.decomposition.incremental_pca import IncrementalPCA
 from cuml.decomposition.pca import PCA
 from cuml.decomposition.tsvd import TruncatedSVD
+from cuml.decomposition.incremental_pca import IncrementalPCA
